Conflicts:
	fortscale/fortscale-domain/src/main/java/fortscale/domain/core/dao/UserActivityRepositoryCustom.java
	fortscale/fortscale-domain/src/main/java/fortscale/domain/core/dao/UserActivityRepositoryImpl.java
	fortscale/fortscale-services/src/main/java/fortscale/services/UserActivityService.java
	fortscale/fortscale-services/src/main/java/fortscale/services/impl/UserActivityServiceImpl.java
	fortscale/fortscale-webapp/src/main/java/fortscale/web/rest/ApiUserActivityController.java
	fortscale/fortscale-webapp/src/main/java/fortscale/web/rest/entities/activity/UserActivityData.java

+++ fortscale/fortscale-collection/src/main/java/fortscale/collection/jobs/activity/UserActivityWorkingHoursHandler.java
@@ -94,7 +94,7 @@ public class UserActivityWorkingHoursHandler extends UserActivityBaseHandler {
 	}
 
 	@Override
-	protected void updateAdditionalActivitySpecificDocumentInDatabase(List<String> dataSources, long currBucketStartTime, long currBucketEndTime, Map<String, Integer> additionalActivityHistogram) {
+	protected void updateAdditionalActivitySpecificDocumentInDatabase(List<String> dataSources, long currBucketStartTime, long currBucketEndTime, Map<String, Double> additionalActivityHistogram) {
 
 	}
 }
